ORM retomado

+++ build.sh
@@ -1,1 +1,4 @@
-mv bin/geleca-framework.swc bin/geleca.3.0.swc
+if [ -f bin/geleca-framework.swc ];
+then
+	mv bin/geleca-framework.swc bin/geleca.3.0.swc
+fi
